# api.yaml
openapi: 3.1.0
info:
  title: NimbusX-API | Content Managing Service
  license:
    name: AGPL 3.0
    url: https://www.gnu.org/licenses/agpl-3.0.txt
  version: 1.0.0
...
security:
  - cookieAuth: [ ]
paths:
...
  /login:
    get:
      tags:
        - "Login"
      responses:
        '200':
          description: Returns data of currently logged in user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
    post:
      tags:
        - "Login"
      security: [ ]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Returns a valid session key
          content:
            application/json:
              schema:
                type: object
                properties:
                  session:
                    type: string

  # User
  /password/{token}:
    parameters:
      - name: token
        in: path
        description: a unique token, which is send to the user via email (included in the reset password link)
        required: true
        schema:
          type: string
    post:
      tags:
        - "User"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Password'
...
  /user/:
    post:
      tags:
        - "User"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
...
  /webhooks/{name}:
    put:
      tags:
        - "Webhooks"
      parameters:
        - name: name
          in: path
          description: Name of the webhook
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              $ref: '#/components/schemas/Webhook'
      responses:
        200:
          description: Update successfully a webhook.
          content:
            application/json:
              schema:
                type: object
                $ref: '#/components/schemas/Webhook'
components:
  schemas:
...
    User:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
    Password:
      type: object
      properties:
        password:
          type: string
...
    Webhook:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        header:
          type: object
        events:
          type: array
          items:
            type: string
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session